+++ sympy/unify/core.py
